refactor(#2): move files around + minor test fixes
