+++ include/omen/libraries/std/stdbool.h
@@ -1,5 +1,5 @@
 #ifndef STDBOOL_H
-#define STDBOOL_h
+#define STDBOOL_H
 
 #define bool _Bool
 #ifndef true
